+++ playground/worker/__tests__/iife/iife-worker.spec.ts
@@ -1,7 +1,15 @@
 import fs from 'node:fs'
 import path from 'node:path'
 import { describe, expect, test } from 'vitest'
-import { isBuild, page, readManifest, testDir, untilUpdated } from '~utils'
+import {
+  isBuild,
+  isServe,
+  page,
+  readManifest,
+  testDir,
+  untilUpdated,
+  viteTestUrl,
+} from '~utils'
 
 test('normal', async () => {
   await untilUpdated(() => page.textContent('.pong'), 'pong')
@@ -18,7 +26,7 @@ test('normal', async () => {
 })
 
 test('named', async () => {
-  await untilUpdated(() => page.textContent('.pong-named'), 'pong', true)
+  await untilUpdated(() => page.textContent('.pong-named'), 'namedWorker', true)
 })
 
 test('TS output', async () => {
@@ -30,7 +38,11 @@ test('inlined', async () => {
 })
 
 test('named inlined', async () => {
-  await untilUpdated(() => page.textContent('.pong-inline-named'), 'pong', true)
+  await untilUpdated(
+    () => page.textContent('.pong-inline-named'),
+    'namedInlineWorker',
+    true,
+  )
 })
 
 test('shared worker', async () => {
@@ -143,3 +155,23 @@ test('import.meta.glob eager in worker', async () => {
     '["',
   )
 })
+
+test.runIf(isServe)('sourcemap boundary', async () => {
+  const response = page.waitForResponse(/my-worker.ts\?type=module&worker_file/)
+  await page.goto(viteTestUrl)
+  const content = await (await response).text()
+  const { mappings } = decodeSourceMapUrl(content)
+  expect(mappings.startsWith(';')).toBeTruthy()
+  expect(mappings.endsWith(';')).toBeFalsy()
+})
+
+function decodeSourceMapUrl(content: string) {
+  return JSON.parse(
+    Buffer.from(
+      content.match(
+        /\/\/[#@]\ssourceMappingURL=\s*data:application\/json;base64,(\S+)/,
+      )?.[1],
+      'base64',
+    ).toString(),
+  )
+}